Add Store Model for divided model , controller
